ft: ZENKO-404 lifecycle service account
Support a service account 'service-lifecycle' for lifecycle feature

* move code related to replication management from lib/management.js
  to extensions/replication/management.js

* add optional callback function 'applyState' that modules can use to
  periodically apply and save their state to management db

* change authentication support for lifecycle to use configuration in
  extensions.lifecycle.auth config, to be able to configure the service
  account name

* update docker-entrypoint.sh with missing S3_HOST and S3_PORT env var
  check

* update docker-entrypoint.sh to add log level support (LOG_LEVEL)
