+++ extensions/lifecycle/LifecycleConfigValidator.js
@@ -4,6 +4,23 @@ const joiSchema = {
     zookeeperPath: joi.string().required(),
     bucketTasksTopic: joi.string().required(),
     objectTasksTopic: joi.string().required(),
+    auth: joi.object({
+        type: joi.alternatives().try('account', 'service', 'vault')
+            .required(),
+        account: joi.string()
+            .when('type', { is: 'account', then: joi.required() })
+            .when('type', { is: 'service', then: joi.required() }),
+        vault: joi.object({
+            host: joi.string().required(),
+            port: joi.number().greater(0).required(),
+            adminPort: joi.number().greater(0)
+                .when('adminCredentialsFile', {
+                    is: joi.exist(),
+                    then: joi.required(),
+                }),
+            adminCredentialsFile: joi.string().optional(),
+        }).when('type', { is: 'vault', then: joi.required() }),
+    }).required(),
     backlogMetrics: {
         zkPath: joi.string().default('/lifecycle/run/backlog-metrics'),
         intervalS: joi.number().default(60),
