+++ extensions/lifecycle/lifecycleProducer/task.js
@@ -1,8 +1,9 @@
 'use strict'; // eslint-disable-line
 const werelogs = require('werelogs');
+const { initManagement } = require('../../../lib/management');
 const LifecycleProducer = require('./LifecycleProducer');
-const { zookeeper, kafka, extensions, s3, auth, transport, log } =
-    require('../../../conf/Config');
+const { zookeeper, kafka, extensions, s3, transport, log } =
+      require('../../../conf/Config');
 
 werelogs.configure({ level: log.logLevel,
                      dump: log.dumpLevel });
@@ -11,9 +12,26 @@ const logger = new werelogs.Logger('Backbeat:Lifecycle:Producer');
 
 const lifecycleProducer =
     new LifecycleProducer(zookeeper, kafka, extensions.lifecycle,
-                          s3, auth, transport);
+                          s3, transport);
 
-lifecycleProducer.start();
+function initAndStart() {
+    initManagement({
+        serviceName: 'lifecycle',
+        serviceAccount: extensions.lifecycle.auth.account,
+    }, error => {
+        if (error) {
+            logger.error('could not load management db',
+                         { error: error.message });
+            setTimeout(initAndStart, 5000);
+            return;
+        }
+        logger.info('management init done');
+
+        lifecycleProducer.start();
+    });
+}
+
+initAndStart();
 
 process.on('SIGTERM', () => {
     logger.info('received SIGTERM, exiting');
