+++ extensions/lifecycle/tasks/LifecycleObjectTask.js
@@ -1,7 +1,10 @@
+const async = require('async');
 const AWS = require('aws-sdk');
 
+const errors = require('arsenal').errors;
 const BackbeatTask = require('../../../lib/tasks/BackbeatTask');
-const async = require('async');
+const { getAccountCredentials } =
+      require('../../../lib/credentials/AccountCredentials');
 const getVaultCredentials =
     require('../../../lib/credentials/getVaultCredentials');
 const { attachReqUids } = require('../../../lib/clients/utils');
@@ -23,17 +26,29 @@ class LifecycleObjectTask extends BackbeatTask {
         this.accountCredsCache = {};
     }
 
-    _getCredentials(canonicalId, cb) {
+    _getCredentials(canonicalId, log, cb) {
         const cachedCreds = this.accountCredsCache[canonicalId];
         if (cachedCreds) {
             return process.nextTick(() => cb(null, cachedCreds));
         }
-        return getVaultCredentials(this.authConfig, canonicalId, 'lifecycle',
-        (err, accountCreds) => cb(err, accountCreds));
+        const credentials = getAccountCredentials(this.lcConfig.auth, log);
+        if (credentials) {
+            this.accountCredsCache[canonicalId] = credentials;
+            return process.nextTick(() => cb(null, credentials));
+        }
+        const { type } = this.lcConfig.auth;
+        if (type === 'vault') {
+            return getVaultCredentials(
+                this.authConfig, canonicalId, 'lifecycle',
+                (err, accountCreds) => cb(err, accountCreds));
+        }
+        return process.nextTick(
+            () => cb(errors.InternalError.customizeDescription(
+                `invalid auth type ${type}`)));
     }
 
     _setupClients(canonicalId, log, done) {
-        this._getCredentials(canonicalId, (err, accountCreds) => {
+        this._getCredentials(canonicalId, log, (err, accountCreds) => {
             if (err) {
                 log.error('error generating new access key', {
                     error: err.message,
