+++ extensions/replication/queueProcessor/task.js
@@ -3,6 +3,7 @@ const werelogs = require('werelogs');
 const QueueProcessor = require('./QueueProcessor');
 const config = require('../../../conf/Config');
 const { initManagement } = require('../../../lib/management');
+const { applyReplicationState } = require('../management');
 
 const zkConfig = config.zookeeper;
 const MetricsProducer = require('../../../lib/MetricsProducer');
@@ -26,7 +27,11 @@ metricsProducer.setupProducer(err => {
         return undefined;
     }
     function initAndStart() {
-        initManagement(error => {
+        initManagement({
+            serviceName: 'replication',
+            serviceAccount: sourceConfig.auth.account,
+            applyState: applyReplicationState,
+        }, error => {
             if (error) {
                 log.error('could not load management db',
                   { error: error.message });
