+++ lib/management.js
@@ -2,7 +2,6 @@ const async = require('async');
 const forge = require('node-forge');
 const arsenal = require('arsenal');
 const config = require('../conf/Config');
-const AWS = require('aws-sdk');
 const werelogs = require('werelogs');
 const bucketclient = require('bucketclient');
 
@@ -18,6 +17,7 @@ const refreshInterval = 5000;
 const Metadata = arsenal.storage.metadata.MetadataWrapper;
 let metadata;
 
+let serviceName;
 let overlayVersion;
 
 // let initialized = false;
@@ -26,7 +26,6 @@ let serviceCredentials = {
 };
 
 const tokenKey = 'auth/zenko/remote-management-token';
-const replicationStateKey = 'configuration/state/replication';
 
 // XXX copy-pasted from S3
 function decryptSecret(instanceCredentials, secret) {
@@ -39,255 +38,21 @@ function decryptSecret(instanceCredentials, secret) {
     });
 }
 
-function saveCurrentReplicationState(db, currentState, cb) {
-    logger.debug('saving replication state', { currentState });
-    metadata.putObjectMD(serviceBucket, replicationStateKey,
-        JSON.stringify(currentState), {}, logger, cb);
-}
-
-function getS3Client(endpoint) {
-    const keys = serviceCredentials.accounts[0].keys;
-    const credentials = new AWS.Credentials(keys.access, keys.secret);
-    const s3Client = new AWS.S3({
-        endpoint,
-        sslEnabled: false,
-        credentials,
-        s3ForcePathStyle: true,
-        signatureVersion: 'v4',
-        httpOptions: { timeout: 0 },
-        maxRetries: 3,
-    });
-    return s3Client;
-}
-
-function ensureBucketExists(stream, end, endpoint, cb) {
-    logger.debug('Checking that bucket exists', { end, endpoint });
-    if (!stream.enabled) {
-        logger.debug('stream is not enabled');
-        return process.nextTick(cb);
-    }
-    const params = {
-        Bucket: end.bucketName,
-    };
-    return getS3Client(endpoint).createBucket(params, err => {
-        if (err && err.code === 'BucketAlreadyOwnedByYou') {
-            return cb();
-        }
-        return cb(err);
-    });
-}
-
-function putVersioning(stream, bucketName, endpoint, cb) {
-    if (!stream.enabled) {
-        return process.nextTick(cb);
-    }
-    const params = {
-        Bucket: bucketName,
-        VersioningConfiguration: {
-            Status: 'Enabled',
-        },
-    };
-    return getS3Client(endpoint).putBucketVersioning(params, err => cb(err));
-}
-
-function installReplicationPolicy(stream, endpoint, cb) {
-    const destinationLocations = stream.destination.locations
-        .map(location => location.name)
-        .join(',');
-    const roleName = 'arn:aws:iam::root:role/s3-replication-role';
-    const params = {
-        Bucket: stream.source.bucketName,
-        ReplicationConfiguration: {
-            Role: `${roleName}`,
-            Rules: [{
-                Destination: {
-                    Bucket: `arn:aws:s3:::${stream.source.bucketName}`,
-                    StorageClass: destinationLocations,
-                },
-                Prefix: stream.source.prefix || '',
-                Status: stream.enabled ? 'Enabled' : 'Disabled',
-            }],
-        },
-    };
-
-    getS3Client(endpoint).putBucketReplication(params, err => {
-        logger.debug('replication applied', {
-            sourceBucket: stream.source.bucketName,
-            destinationLocations, error: err });
-        if (err && err.code === 'NoSuchBucket' && !stream.enabled) {
-            return cb();
-        }
-        return cb(err);
-    });
-}
-
-function putReplication(stream, cb) {
-    logger.debug('Changing replication for stream to stream.enabled',
-        { streamId: stream.streamId });
-    const cfg = config.extensions.replication.source.s3;
-    const endpoint = `${cfg.host}:${cfg.port}`;
-
-    async.waterfall([
-        done => ensureBucketExists(stream, stream.source, endpoint, done),
-        done => putVersioning(stream, stream.source.bucketName, endpoint, done),
-        // TODO add service account in source & target bucket ACLs
-        done => installReplicationPolicy(stream, endpoint, done),
-    ], cb);
-}
-
-/* eslint-disable no-unused-vars */
-function disableReplication(stream, cb) {
-    logger.debug('disabling replication for stream',
-        { streamId: stream.streamId });
-    return cb(null, stream);
-}
-/* eslint-enable no-unused-vars */
-
-function applyReplicationState(conf, currentStateSerialized, db, cb) {
-    if (!conf.replicationStreams) {
-        return process.nextTick(cb);
-    }
-
-    const currentState = currentStateSerialized ?
-        JSON.parse(currentStateSerialized) : {
-            streams: {},
-            overlayVersion: 0,
-        };
-
-    if (currentState.overlayVersion >= conf.version) {
-        return process.nextTick(cb);
-    }
-
-    function streamConfigured(stream, cb) {
-        currentState.streams[stream.streamId] = stream;
-        saveCurrentReplicationState(db, currentState, cb);
-    }
-
-    function deleteStream(streamId, cb) {
-        logger.debug('deleting obsolete stream`', { streamId });
-        const stream = currentState.streams[streamId];
-        stream.enabled = false;
-        putReplication(stream, err => {
-            if (err) {
-                return cb(err);
-            }
-            delete currentState.streams[streamId];
-            return saveCurrentReplicationState(db, currentState, cb);
-        });
-    }
-
-    function arraysEqual(arr1, arr2) {
-        if (arr1.length !== arr2.length) {
-            return false;
-        }
-        for (let i = arr1.length; i--;) {
-            if (arr1[i] !== arr2[i]) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    function putReplicationAndStreamConfigured(stream, cb) {
-        putReplication(stream, err => {
-            if (err) {
-                return cb(err);
-            }
-            return streamConfigured(stream, cb);
-        });
-    }
-
-    function configureStream(stream, cb) {
-        if (currentState.streams[stream.streamId]) {
-            if (!currentState.streams[stream.streamId].enabled
-                && stream.enabled) {
-                logger.debug('enabling existing stream', {
-                    streamId: stream.streamId });
-                putReplicationAndStreamConfigured(stream, cb);
-            } else if (currentState.streams[stream.streamId].enabled
-                && !stream.enabled) {
-                logger.debug('disabling existing stream',
-                    { streamId: stream.streamId });
-                putReplicationAndStreamConfigured(stream, cb);
-            } else if (currentState.streams[stream.streamId].source.prefix
-                !== stream.source.prefix) {
-                logger.debug('changing stream source prefix',
-                    { streamId: stream.streamId });
-                putReplicationAndStreamConfigured(stream, cb);
-            } else if (!arraysEqual(currentState.streams[stream.streamId]
-                .destination.locations, stream.destination.locations)) {
-                logger.debug('changing stream destination locations',
-                    { streamId: stream.streamId });
-                putReplicationAndStreamConfigured(stream, cb);
-            } else {
-                logger.debug('stream has expected state, doing nothing',
-                { streamId: stream.streamId, streamEnabled: stream.enabled });
-                return cb();
-            }
-        } else {
-            if (stream.enabled) {
-                logger.debug('enabling new stream',
-                    { streamId: stream.streamId });
-                putReplicationAndStreamConfigured(stream, cb);
-            } else {
-                logger.debug('new disabled stream, doing nothing',
-                    { streamId: stream.streamId });
-                return cb();
-            }
-        }
-        return undefined;
-    }
-
-    const existingStreamIds = Object.keys(currentState.streams);
-    const incomingConfigStreamIds =
-        conf.replicationStreams.map(s => s.streamId);
-    const deletedStreamIds =
-        existingStreamIds.filter(s => !incomingConfigStreamIds.includes(s));
-
-    // TODO group streams per source bucket
-    async.eachSeries(conf.replicationStreams, configureStream, err => {
-        if (err) {
-            return cb(err);
-        }
-        currentState.overlayVersion = conf.version;
-        async.eachSeries(deletedStreamIds, deleteStream, err => {
-            if (err) {
-                return cb(err);
-            }
-            return saveCurrentReplicationState(db, currentState, cb);
-        });
-        return undefined;
-    });
-    return undefined;
-}
-
-function saveServiceCredentials(conf, auth) {
-    // const instanceAuth = JSON.parse(auth);
-    const instanceAuth = auth;
-    const serviceAccount = (conf.users || []).find(
-        u => u.accountType === 'service-replication');
-    if (!serviceAccount) {
-        return;
-    }
-
-    // TODO generate a proper canonicalID during the service account /
-    // instance provisioning
+function saveServiceCredentials(conf, params, instanceAuth) {
     // TODO use a proper account id in arn
-
-    const secretKey = decryptSecret(instanceAuth, serviceAccount.secretKey);
     serviceCredentials = {
-        accounts: [{
-            name: 'service-replication',
-            arn: 'aws::iam:234456789012:root',
-            // canonicalID:
-// '79a59df900b949e55d96a1e698fbacedfd6e09d98eacf8f8d5218e7cd47ef2be',
-            canonicalID: '12349df900b949e55d96a1e698fbacedfd6e09d98eacf8f8d52',
-            displayName: serviceAccount.userName,
-            keys: {
-                access: serviceAccount.accessKey,
-                secret: secretKey,
-            },
-        }],
+        accounts: (conf.users || [])
+            .filter(u => u.accountType === params.serviceAccount)
+            .map(u => ({
+                name: u.accountType,
+                arn: 'aws::iam:234456789012:root',
+                canonicalID: u.canonicalId,
+                displayName: u.userName,
+                keys: {
+                    access: u.accessKey,
+                    secret: decryptSecret(instanceAuth, u.secretKey),
+                },
+            })),
     };
     // initialized = true;
 }
@@ -332,13 +97,57 @@ function patchConfiguration(conf, cb) {
     return process.nextTick(cb, null, conf);
 }
 
-function initManagement(done) {
+function _getExtensionStateKey() {
+    return `configuration/state/${serviceName}`;
+}
+
+function _loadExtensionState(cb) {
+    logger.debug(`loading ${serviceName} state`);
+    metadata.getObjectMD(
+        serviceBucket, _getExtensionStateKey(),
+        {}, logger, (err, currentState) => {
+            if (err && err.NoSuchKey) {
+                /* eslint-disable no-param-reassign */
+                return cb(null, null);
+            }
+            if (err) {
+                return cb(err);
+            }
+            return cb(null, currentState);
+        });
+}
+
+function _refreshExtensionState(conf, params, done) {
+    if (!params.applyState) {
+        return process.nextTick(done);
+    }
+    return async.waterfall([
+        cb => _loadExtensionState(cb),
+        (currentState, cb) => params.applyState(conf, currentState, cb),
+    ], done);
+}
+
+/**
+ * Initialize Orbit management layer
+ *
+ * @param {object} params - params object
+ * @param {string} params.serviceName - name of service to manage
+ * @param {string} [params.serviceAccount] - name of managed service
+ *   account, if any
+ * @param {function} [params.applyState] - callback function to
+ *   apply extension-specific state set from management layer
+ * @param {function} done - callback function when init is complete
+ * @return {undefined}
+ */
+function initManagement(params, done) {
     if (process.env.REMOTE_MANAGEMENT_DISABLE &&
         process.env.REMOTE_MANAGEMENT_DISABLE !== '0') {
         logger.info('remote management disabled');
         return process.nextTick(done);
     }
-    const params = {
+    serviceName = params.serviceName;
+
+    const mdParams = {
         bucketdBootstrap: ['localhost'],
         bucketdLog: null,
         https: null,
@@ -371,7 +180,7 @@ function initManagement(done) {
             cb => {
                 if (!setup) {
                     setup = true;
-                    metadata = new Metadata('mongodb', params, bucketclient,
+                    metadata = new Metadata('mongodb', mdParams, bucketclient,
                         logger);
                     return metadata.setup(() => cb());
                 }
@@ -389,20 +198,10 @@ function initManagement(done) {
                 if (err) {
                     return cb(err);
                 }
-                saveServiceCredentials(conf, instanceAuth);
+                saveServiceCredentials(conf, params, instanceAuth);
                 return cb(null, conf);
             }),
-            (conf, cb) => metadata.getObjectMD(serviceBucket,
-            replicationStateKey, {}, logger, (err, currentReplicationState) => {
-                if (err && err.NoSuchKey) {
-                    /* eslint-disable no-param-reassign */
-                    currentReplicationState = null;
-                } else if (err) {
-                    return cb(err);
-                }
-                return applyReplicationState(conf, currentReplicationState,
-                    metadata, cb);
-            }),
+            (conf, cb) => _refreshExtensionState(conf, params, cb),
         ], done);
     }
 
@@ -424,7 +223,16 @@ function getLatestServiceAccountCredentials() {
     return serviceCredentials;
 }
 
+function saveExtensionState(currentState, cb) {
+    logger.debug(`saving ${serviceName} state`, { currentState });
+    metadata.putObjectMD(
+        serviceBucket, _getExtensionStateKey(),
+        JSON.stringify(currentState), {}, logger, cb);
+}
+
+
 module.exports = {
     initManagement,
     getLatestServiceAccountCredentials,
+    saveExtensionState,
 };
